refactor(website): Break App into App, ContextWrapper, and PageRouter
Split the monolithic `App` component into smaller subcomponents to handle different parts of the website.

In particular, broke off the `ContextWrapper` to contain all of the required Context Providers, as those require a lot of nesting that doesn't need to be in `App` as they are all effectively passthroughs.

Also broke off the `PageRouter` to handle top-level React Routing.

Closes #35.