feat(website): Initialize Data for the Website (#37)
* feat(website): Define the Typescript Definitions for Data Objects

Working on #6, #9, #12, #15, #18.

* feat(website): Define Local Cache for the Website Data

This provides a performance boost since once the data is loaded from the API once, we don't need to do any more reads. This saves billing time for the back end, Firestore calls, and latency on the front end.

Working on #6, #9, #12, #15, #18.

* feat(website): Define the GraphQL Queries for Loading the Initial Data

Working on #6, #9, #12, #15, #18.

* feat(website): Define the React Components to Initialize the Local Cache

All data that needs to be listed from the UI is now loaded to the website.

Working on #6, #9, #12, #15, #18.